requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pypi_0
anyio=2.2.0=py36h5fab9bb_0
astunparse=1.6.3=pypi_0
async_generator=1.10=py_0
attrs=20.3.0=pyhd3deb0d_0
babel=2.9.0=pyhd3deb0d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.3=pyhd8ed1ab_0
bleach=3.3.0=pyh44b312d_0
ca-certificates=2020.12.5=ha878542_0
cachetools=4.2.1=pypi_0
certifi=2020.12.5=py36h5fab9bb_1
chardet=4.0.0=pypi_0
contextvars=2.4=py_0
cycler=0.10.0=pypi_0
dataclasses=0.8=pyh787bdff_0
decorator=5.0.6=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
gast=0.3.3=pypi_0
google-auth=1.28.1=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.37.0=pypi_0
h5py=2.10.0=pypi_0
idna=2.10=pypi_0
imagehash=4.2.0=pypi_0
immutables=0.14=py36h8c4c3a4_1
importlib-metadata=3.10.0=py36h5fab9bb_0
ipykernel=5.5.3=py36hcb3619a_0
ipython=7.16.1=py36he448a4c_2
ipython_genutils=0.2.0=py_1
jedi=0.17.2=py36h5fab9bb_1
jinja2=2.11.3=pyh44b312d_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter-packaging=0.7.12=pyhd8ed1ab_0
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_core=4.7.1=py36h5fab9bb_0
jupyter_server=1.4.1=py36h5fab9bb_0
jupyterlab=3.0.13=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.4.0=pyhd8ed1ab_0
jupyterthemes=0.20.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
lesscpy=0.14.0=pypi_0
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.1.0=hdf63c60_0
markdown=3.3.4=pypi_0
markupsafe=1.1.1=py36he6145b8_2
matplotlib=3.3.4=pypi_0
mistune=0.8.4=py36h1d69622_1002
nbclassic=0.2.6=pyhd8ed1ab_0
nbclient=0.5.3=pyhd8ed1ab_0
nbconvert=6.0.7=py36h5fab9bb_3
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd8ed1ab_0
notebook=6.0.3=py36h9f0ad1d_1
numpy=1.18.5=pypi_0
oauthlib=3.1.0=pypi_0
openssl=1.1.1k=h27cfd23_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pyh44b312d_0
pandas=1.1.5=pypi_0
pandoc=2.12=h7f98852_0
pandocfilters=1.4.2=py_1
parso=0.7.1=pyh9f0ad1d_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.2.0=pypi_0
pip=21.0.1=py36h06a4308_0
ply=3.11=pypi_0
prometheus_client=0.10.1=pyhd8ed1ab_0
prompt-toolkit=3.0.18=pyha770c72_0
protobuf=3.15.8=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pygments=2.8.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.17.3=py36h1d69622_1
python=3.6.13=hdb3f193_0
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytz=2021.1=pyhd8ed1ab_0
pywavelets=1.1.1=pypi_0
pyzmq=20.0.0=py36h2b1bd32_1
readline=8.1=h27cfd23_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scipy=1.4.1=pypi_0
send2trash=1.5.0=py_0
setuptools=52.0.0=py36h06a4308_0
six=1.15.0=pyh9f0ad1d_0
sniffio=1.2.0=py36h5fab9bb_1
sqlite=3.35.4=hdfb4753_0
tensorboard=2.4.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow-estimator=2.3.0=pypi_0
tensorflow-gpu=2.3.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.4=py36h5fab9bb_0
testpath=0.4.4=py_0
tk=8.6.10=hbc83047_0
tornado=6.1=py36h1d69622_0
tqdm=4.60.0=pypi_0
traitlets=4.3.3=py36h9f0ad1d_1
typing_extensions=3.7.4.3=py_0
urllib3=1.26.4=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.3=h58526e2_3
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
